requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-autoslug==1.9.9
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
phonenumbers==8.13.18
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sender==0.3
sqlparse==0.4.4
twilio==8.5.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
